Исправить иерархию контроллеров. Удалить контроллеры Action и Chest. Создать контроллеры Player и Room. Перераспределить эндпоинты.
