+++ Controllers/GameController.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using TextGame;
-
 namespace TextGame.Controllers
 {
     [ApiController]
@@ -14,7 +12,7 @@ namespace TextGame.Controllers
             this.gameRepository = gameRepository;
         }
 
-        [HttpPost("start")]
+        [HttpPost("")]
         public IResult Start()
         {
             try
@@ -28,47 +26,7 @@ namespace TextGame.Controllers
             }
         }
 
-        [HttpGet("currentroom")]
-        public IResult ShowCurrentRoom()
-        {
-            try
-            {
-                var room = gameRepository.ShowCurrentRoom();
-                return Results.Ok(room);
-            }
-            catch (Exception ex)
-            {
-                return Results.Problem($"Ошибка при получении текущей комнаты: {ex.Message}");
-            }
-        }
-
-        [HttpGet("inventory")]
-        public IResult ShowInventory()
-        {
-            try
-            {
-                var inventory = gameRepository.ShowInventory();
-                return Results.Ok(inventory);
-            }
-            catch (Exception ex)
-            {
-                return Results.Problem($"Ошибка при получении инвентаря: {ex.Message}");
-            }
-        }
-
-        [HttpGet("coins")]
-        public IResult ShowCoins()
-        {
-            try
-            {
-                var coins = gameRepository.ShowCoins();
-                return Results.Ok(coins);
-            }
-            catch (Exception ex)
-            {
-                return Results.Problem($"Ошибка при получении количества монет: {ex.Message}");
-            }
-        }
+        
 
     }
 }
